Revert ABI fixed width types (#773)
Reverts villagesql/villagesql-server#755

It seems we should investigate this more, perhaps when we tackle
Windows.

We need to understand whether this is actually a problem or if this is
unnecessary.

Created by bonsai from 29157a5e440438ab72a49079acd554b3fd5cba9f

+++ villagesql/services/sys_var_access.h
@@ -17,7 +17,7 @@
 #ifndef VILLAGESQL_SERVICES_SYS_VAR_ACCESS_H
 #define VILLAGESQL_SERVICES_SYS_VAR_ACCESS_H
 
-#include <cstdint>
+#include <stddef.h>
 #include <string_view>
 
 namespace villagesql::services {
@@ -55,7 +55,7 @@ void unregister_one_sys_var(std::string_view extension_name,
 //
 // Returns false on success, true on error.
 bool get_variable(const char *component_name, const char *name, void **val,
-                  uint64_t *val_len);
+                  size_t *val_len);
 
 // Set the value of a system variable.
 //
